Improve display of QAM lineups during configure.
The data for QAM lineups does not seem to include valid 'name' and
'location' fields. These fields are either missing if it is the first
lineup or seem to erroneously contain the values from the previous lineup.

In the first case you would get errors such as 'Use of uninitialized value
in concatenation (.) or string at ./tv_grab_na_sd_json line 351, <STDIN>
line 10'. In the second case incorrect information was displayed.