If --days is not specified use the value from the config file.
When --days is not specified I think it makes more sense to use the days
value from the config file instead of the default value of 5 in the
ParseOptions function.

For example, mythtv does not pass the --days option when the preferred
method is allatonce. If we used the default value of 5 mythtv would never
grab more than 5 days of data.