Improve error handling, tweak status messages.
The sd_json_request function now dies on most errors. We now try to create
a more useful error message instead of dumping the full response object.
In some cases we use try/catch around the sd_json_request call to recover
from errors.

There is probably still a lot of improvements that can be made to error
checking and handling but I think this covers the most common cases.