Make sure lineups in configuration are enabled for the account in use.
If a lineup is enabled in your grabber configuration which is not enabled
for your account an appropriate error message is now displayed and the
process aborts. Previously this condition could go unnoticed with
incomplete results from the cache. If the lineup information was
not cached it would error out with no clear indication of why.