Update to use 20141201 API
So far this has had minimal testing but since data updates for the old API
are being discontinued it is better than nothing. It at least works for
the OTA listing I actually use on a regular basis. I hope to do more
testing soon especially with other lineup types.