Only use lineups from current config when building channel/map index.
If a station exists in more than one lineup there could potentially be
more than one channel mapping (channel number). Previously we looked
through all lineups enabled for the account to find channel mappings which
could result in using a mapping from a lineup that is not even enabled in
the current config.