Allow configuration of previously-shown format.
By the XMLTV spec the time should not be included (unless it is actually
known). Older versions of MythTV require it otherwise MythTV assumes UTC
which causes the wrong date to be stored. However there have been reports
that including the time breaks newer versions of MythTV. Lets make it
configurable to accommodate both cases.